devtools/server/tests/mochitest/test_memory_attach_01.html

<!DOCTYPE HTML>
<html>
<!--
Bug 960671 - Test attaching and detaching from a memory actor.
-->
<head>
  <meta charset="utf-8">
  <title>Memory monitoring actor test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="memory-helpers.js" type="application/javascript"></script>
<script>
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  (async function() {
    const { memory, client } = await startServerAndGetSelectedTabMemory();
    await memory.attach();
    ok(true, "Shouldn't have gotten an error attaching.");
    await memory.detach();
    ok(true, "Shouldn't have gotten an error detaching.");
    destroyServerAndFinish(client);
  })();
};
</script>
</pre>
</body>
</html>
